Refactored package structure for proper multi-module build. Added TOML module. Fixed some things in AbstractConfigCodec; subclassing is now easier.
