--- build.gradle
@@ -1,57 +1,0 @@
-//file:noinspection GroovyAssignabilityCheck
-plugins {
-    id 'java'
-    id 'maven-publish'
-}
-
-group 'com.github.steank'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    //annotations
-    compileOnly "org.jetbrains:annotations:${project.jetbrains_annotations_version}"
-
-    //use junit for unit testing
-    testImplementation "org.junit.jupiter:junit-jupiter-api:${project.junit_version}"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:${project.junit_version}"
-
-    //use mockito for mocking
-    testImplementation "org.mockito:mockito-core:${project.mockito_version}"
-
-    //annotations for tests
-    testCompileOnly "org.jetbrains:annotations:${project.jetbrains_annotations_version}"
-}
-
-java {
-    withSourcesJar()
-    withJavadocJar()
-}
-
-test {
-    useJUnitPlatform()
-}
-
-publishing {
-    publications {
-        myPublication(MavenPublication) {
-            from components.java
-        }
-    }
-
-    repositories {
-        maven {
-            name = "ethylene"
-            url = "https://maven.cloudsmith.io/steank-f1g/ethylene/"
-
-            credentials {
-                //used to publish locally
-                username = System.getenv("CLOUDSMITH_USERNAME")
-                password = System.getenv("CLOUDSMITH_PASSWORD")
-            }
-        }
-    }
-}
