+++ ethylene-core/build.gradle
@@ -1,0 +1,56 @@
+//file:noinspection GroovyAssignabilityCheck
+plugins {
+    id 'java-library'
+    id 'maven-publish'
+}
+
+group 'com.github.steanky'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    //annotations
+    compileOnlyApi "org.jetbrains:annotations:${project.jetbrains_annotations_version}"
+
+    //use junit for unit testing
+    testImplementation "org.junit.jupiter:junit-jupiter-api:${project.junit_version}"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:${project.junit_version}"
+
+    //use mockito for mocking
+    testImplementation "org.mockito:mockito-core:${project.mockito_version}"
+
+    //annotations for tests
+    testCompileOnly "org.jetbrains:annotations:${project.jetbrains_annotations_version}"
+}
+
+java {
+    withSourcesJar()
+    withJavadocJar()
+}
+
+test {
+    useJUnitPlatform()
+}
+
+publishing {
+    publications {
+        myPublication(MavenPublication) {
+            from components.java
+        }
+    }
+
+    repositories {
+        maven {
+            name = "ethylene"
+            url = "https://maven.cloudsmith.io/steank-f1g/ethylene/"
+
+            credentials {
+                username = System.getenv("CLOUDSMITH_USERNAME")
+                password = System.getenv("CLOUDSMITH_PASSWORD")
+            }
+        }
+    }
+}
