+++ ethylene-toml/build.gradle
@@ -1,0 +1,31 @@
+plugins {
+    id 'java-library'
+}
+
+group 'com.github.steanky'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    api project(':ethylene-core')
+
+    //toml parser
+    implementation 'com.moandjiezana.toml:toml4j:0.7.2'
+
+    //use junit for unit testing
+    testImplementation "org.junit.jupiter:junit-jupiter-api:${project.junit_version}"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:${project.junit_version}"
+
+    //use mockito for mocking
+    testImplementation "org.mockito:mockito-core:${project.mockito_version}"
+
+    //annotations for tests
+    testCompileOnly "org.jetbrains:annotations:${project.jetbrains_annotations_version}"
+}
+
+test {
+    useJUnitPlatform()
+}
