+++ ethylene-toml/src/test/java/com/github/steanky/ethylene/toml/TomlCodecTest.java
@@ -1,0 +1,7 @@
+package com.github.steanky.ethylene.toml;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TomlCodecTest {
+
+}
