+++ settings.gradle
@@ -1,4 +1,5 @@
 rootProject.name = 'ethylene'
-
+include 'ethylene-core'
+include 'ethylene-toml'
